Initial code structure
This initial code structure is based on the work of David Sand, which is inspired by the work on SUAVE (Stanford/Embraer code) with some modifications:

The code is inside the /src folder, with the setup.py, README, LICENSE and other generic files on top-level. This structure allows ADR to be installed as a Python Package throught pip.