Oh fuck, really big commit
Ok, so this is a big change. Most of the code structure was improved in order to better answer for the needs of the modules.

The Takeoff code is functional and uses almost 100% of it's data from objects instantiation, like it should be.

The main.py now contains an example of plane instantiation and use of it's attributes.

Shit is getting real.

VUASHHHHHHHHHHHHHHHHHH