Alpha ranges implemented
Ok, don't kill me for the size of the commit. The fact is: everybody had to change.

Min e Max alphas for power analysis and for trimm analysis are now implemented and being used in a lot of places.

We can also set an alpha for the plane and it auto updates it's aerodynamic surfaces, trimmed!

Everything working again.