TrimmRangeChecker implemented
This is the first of a series of  Checkers. Those checkers are instantiated and called by the Main to check if a series of conditions are met.

This one checks if the tail can trimm the airplane for the entire range of alphas that the plane can experience in flight.