Add new ADR run method
Now there is an special module called insert_genes, which is called when genes are passed, and modifies the original parameters dictionary for a new modified one (based on the genes).

There's also a who_you_gonna_call method which has 3 variables: the genes, use_genes and use_own_parameters. This is the new straughtforward way for anybody to run ADR.

There's no more parameters_optmizer also. Now the original parameters dictionary is modified by the insert_genes method.