Change local repo name from avl to avl_repo
If the name of the repo stays as "avl", when AVLWrapper tries to execute "avl" it will fing the repo folder "avl" instead of the the "avl" executable on the bin subfolder. This will them give a PermissionDenied error as the program will be trying to execute a folder.

It took me a total of three hours and a night of bad sleep dreaming with spiders to figure that out.

Fuck.